Fix NullReferenceException
A NullReferenceException seems to have happened at some point. I don't know specifically what as null or why, but a hunch is that model binding to Options failed for some reason, and then the view didn't handle it properly. This commit fixes that by doing the null checking in MetaModel consistently.